feat: supporting reduce with no initializer
Arrays support calling reduce with no initializer, which makes the first call having the first and
second elements, and the subsequent calls to behave as any other reduction call.
This implementation makes reduce on fluent to do behave just like that, but with a stronger
type definition.

This type of call is ideal to be used when the result have the same type of the elements.
If you want to change the result type, you can excplicitly inform it through generics, but be aware:
if the array have only one element, the result will still have the same type of the elements,
as the callback won't be called. For this reason, the result type in this case is ambiguous, and
is up to the lib user to guarantee its consistency